Add missing hooks and remove legacy services
